refactor: isolate Listbox stories
