+++ packages/react-aria-components/stories/ListBox.stories.tsx
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2022 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {Header, ListBox, Section, Separator, Text} from 'react-aria-components';
+import {MyListBoxItem} from './utils';
+import React from 'react';
+import styles from '../example/index.css';
+
+export default {
+  title: 'React Aria Components'
+};
+
+export const ListBoxExample = (args) => (
+  <ListBox className={styles.menu} {...args} aria-label="test listbox">
+    <MyListBoxItem>Foo</MyListBoxItem>
+    <MyListBoxItem>Bar</MyListBoxItem>
+    <MyListBoxItem>Baz</MyListBoxItem>
+    <MyListBoxItem href="http://google.com">Google</MyListBoxItem>
+  </ListBox>
+);
+
+ListBoxExample.story = {
+  args: {
+    selectionMode: 'none',
+    selectionBehavior: 'toggle'
+  },
+  argTypes: {
+    selectionMode: {
+      control: {
+        type: 'radio',
+        options: ['none', 'single', 'multiple']
+      }
+    },
+    selectionBehavior: {
+      control: {
+        type: 'radio',
+        options: ['toggle', 'replace']
+      }
+    }
+  }
+};
+
+// Known accessibility false positive: https://github.com/adobe/react-spectrum/wiki/Known-accessibility-false-positives#listbox
+// also has a aXe landmark error, not sure what it means
+export const ListBoxSections = () => (
+  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="test listbox with section">
+    <Section className={styles.group}>
+      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
+      <MyListBoxItem>Foo</MyListBoxItem>
+      <MyListBoxItem>Bar</MyListBoxItem>
+      <MyListBoxItem>Baz</MyListBoxItem>
+    </Section>
+    <Separator style={{borderTop: '1px solid gray', margin: '2px 5px'}} />
+    <Section className={styles.group}>
+      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
+      <MyListBoxItem>Foo</MyListBoxItem>
+      <MyListBoxItem>Bar</MyListBoxItem>
+      <MyListBoxItem>Baz</MyListBoxItem>
+    </Section>
+  </ListBox>
+);
+
+export const ListBoxComplex = () => (
+  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="listbox complex">
+    <MyListBoxItem>
+      <Text slot="label">Item 1</Text>
+      <Text slot="description">Description</Text>
+    </MyListBoxItem>
+    <MyListBoxItem>
+      <Text slot="label">Item 2</Text>
+      <Text slot="description">Description</Text>
+    </MyListBoxItem>
+    <MyListBoxItem>
+      <Text slot="label">Item 3</Text>
+      <Text slot="description">Description</Text>
+    </MyListBoxItem>
+  </ListBox>
+);
