+++ packages/react-aria-components/stories/index.stories.tsx
@@ -25,73 +25,6 @@ export default {
   title: 'React Aria Components'
 };
 
-export const ListBoxExample = (args) => (
-  <ListBox className={styles.menu} {...args} aria-label="test listbox">
-    <MyListBoxItem>Foo</MyListBoxItem>
-    <MyListBoxItem>Bar</MyListBoxItem>
-    <MyListBoxItem>Baz</MyListBoxItem>
-    <MyListBoxItem href="http://google.com">Google</MyListBoxItem>
-  </ListBox>
-);
-
-ListBoxExample.story = {
-  args: {
-    selectionMode: 'none',
-    selectionBehavior: 'toggle'
-  },
-  argTypes: {
-    selectionMode: {
-      control: {
-        type: 'radio',
-        options: ['none', 'single', 'multiple']
-      }
-    },
-    selectionBehavior: {
-      control: {
-        type: 'radio',
-        options: ['toggle', 'replace']
-      }
-    }
-  }
-};
-
-// Known accessibility false positive: https://github.com/adobe/react-spectrum/wiki/Known-accessibility-false-positives#listbox
-// also has a aXe landmark error, not sure what it means
-export const ListBoxSections = () => (
-  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="test listbox with section">
-    <Section className={styles.group}>
-      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
-      <MyListBoxItem>Foo</MyListBoxItem>
-      <MyListBoxItem>Bar</MyListBoxItem>
-      <MyListBoxItem>Baz</MyListBoxItem>
-    </Section>
-    <Separator style={{borderTop: '1px solid gray', margin: '2px 5px'}} />
-    <Section className={styles.group}>
-      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
-      <MyListBoxItem>Foo</MyListBoxItem>
-      <MyListBoxItem>Bar</MyListBoxItem>
-      <MyListBoxItem>Baz</MyListBoxItem>
-    </Section>
-  </ListBox>
-);
-
-export const ListBoxComplex = () => (
-  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="listbox complex">
-    <MyListBoxItem>
-      <Text slot="label">Item 1</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-    <MyListBoxItem>
-      <Text slot="label">Item 2</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-    <MyListBoxItem>
-      <Text slot="label">Item 3</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-  </ListBox>
-);
-
 export const TagGroupExample = (props) => (
   <TagGroup {...props}>
     <Label>Categories</Label>
